Split some shared fields into new files.
The DATA filename lists the records that use it, because there are
records which use DATA fields with different structures. If any of the
other fields with separate files are found to be in the same position,
they will need renaming. A repository search should show up all the
links that would need updating.