Update CHAL.md
Added the new ICON and MICO subrecords. Tried to add the description for Type Values, but the table is too wide and the end result looks bad in browser. Seems that github theme does not take into account a large screen and persist into displaying only on half of the window, even when editing/viewing in full screen.